+++ test/bot.test.js
@@ -9,50 +9,39 @@ import model, {Model} from '../src/model/model';
 import {deserialize, serialize, createModelSchema, ref, list, child} from 'serializr';
 import botFactory from '../src/factory/botFactory';
 import roster from '../src/store/xmpp/rosterService';
-import Bot, {VISIBILITY_FRIENDS} from '../src/model/Bot';
+import Bot, {LOCATION, VISIBILITY_FRIENDS} from '../src/model/Bot';
 import profile from '../src/store/profileStore';
 
 let botData;
-let user, password, server;
+let user, password, server, botId;
 let friend;
 describe("bot", function() {
-  // step("expect title", async function(done){
-  //   try {
-  //     const data = testDataNew(11);
-  //     const {user, password, server} = await xmpp.register(data.resource, data.provider_data);
-  //     const logged = await xmpp.connect(user, password, server);
-  //     await bot.create({});
-  //     await xmpp.disconnect(null);
-  //     done("title should be required");
-  //   } catch (e){
-  //     done()
-  //   }
-  // });
-  // step("expect radius", async function(done){
-  //   try {
-  //     const data = testDataNew(11);
-  //     const {user, password, server} = await xmpp.register(data.resource, data.provider_data);
-  //     const logged = await xmpp.connect(user, password, server);
-  //     await bot.create({title:'Bot title'});
-  //     await xmpp.disconnect(null);
-  //     done("radius should be required");
-  //   } catch (e){
-  //     done()
-  //   }
-  // });
-  // step("expect server side error about short title", async function(done){
-  //   try {
-  //     const data = testDataNew(11);
-  //     const {user, password, server} = await xmpp.register(data.resource, data.provider_data);
-  //     const logged = await xmpp.connect(user, password, server);
-  //     await bot.create({title:'Bot title', radius:10, location: {latitude:11.1, longitude:12.5, accuracy:2}});
-  //     await xmpp.disconnect(null);
-  //     done('error should be thrown');
-  //   } catch (e){
-  //     expect(e).to.be.equal('Missing shortnamefield')
-  //     done()
-  //   }
-  // });
+  step("generate id", async function(done){
+    try {
+      const data = testDataNew(11);
+      const {user, password, server} = await xmpp.register(data.resource, data.provider_data);
+      const logged = await xmpp.connect(user, password, server);
+      botStore.create();
+      when(()=>botStore.bot.id, () => {
+        expect(botStore.bot.id).to.be.not.undefined;
+        done();
+      });
+    } catch (e){
+      done(e);
+    }
+  });
+  step("expect title", async function(done){
+    try {
+      const data = testDataNew(11);
+      const {user, password, server} = await xmpp.register(data.resource, data.provider_data);
+      const logged = await xmpp.connect(user, password, server);
+      await bot.create({id: botId});
+      await xmpp.disconnect(null);
+      done("title should be required");
+    } catch (e){
+      done()
+    }
+  });
   
   step("register/login friend", async function(done){
     const data = testDataNew(12);
@@ -62,7 +51,7 @@ describe("bot", function() {
     await xmpp.disconnect(null);
     done();
   });
-  step("expect creation", async function(done) {
+  step("add friend", async function(done) {
     try {
       const data = testDataNew(11);
       const shortname = undefined;
@@ -73,17 +62,16 @@ describe("bot", function() {
       server = response.server;
       image = 'testimage';
       const logged = await xmpp.connect(user, password, server);
-    
+
       // add friend
       roster.subscribe(friend);
       await roster.add({user: friend});
-      await xmpp.disconnect(null);
+      await xmpp.disconnect();
       done()
     } catch (e) {
       done(e)
     }
   });
-  
   step("register/login friend and confirm add friend", async function(done){
     const data = testDataNew(12);
     const {user, password, server} = await xmpp.register(data.resource, data.provider_data);
@@ -107,18 +95,30 @@ describe("bot", function() {
       image = 'testimage';
       const logged = await xmpp.connect(user, password, server);
       roster.authorize(friend);
-      
-      res = await bot.create({type:'location', title:'Bot title', isNew: true, radius:10, shortname, description, image,
-        location: {latitude:11.1, longitude:12.5, accuracy:2}, visibility: VISIBILITY_FRIENDS, newAffiliates:[{user:friend}]});
-      expect(res.id).to.be.not.undefined;
-      expect(res.server).to.be.not.undefined;
-      expect(res.title).to.be.equal('Bot title');
-      expect(res.shortname).to.be.equal(shortname);
-      expect(res.description).to.be.equal(description);
-      expect(res.image).to.be.equal(image);
-      botData = res;
-      await xmpp.disconnect(null);
-      done();
+
+      botStore.create({type:'location', title:'Bot title', radius:10, shortname, description,
+        location: {latitude:11.1, longitude:12.5, accuracy:2}, image, visibility: VISIBILITY_FRIENDS, newAffiliates:[{user:friend}]});
+  
+      when (()=>botStore.bot.id, async ()=>{
+        try {
+          expect(botStore.bot.id).to.be.not.undefined;
+          botId = botStore.bot.id;
+          await botStore.save();
+          const res = botStore.bot;
+          expect(res.id).to.be.equal(botId);
+          expect(res.server).to.be.not.undefined;
+          expect(res.title).to.be.equal('Bot title');
+          expect(res.shortname).to.be.equal(shortname);
+          expect(res.description).to.be.equal(description);
+          expect(res.image.id).to.be.equal(image);
+          botData = res;
+          await xmpp.disconnect(null);
+          done();
+  
+        } catch (e){
+          done(e)
+        }
+      });
     } catch (e){
       done(e)
     }
@@ -126,21 +126,21 @@ describe("bot", function() {
   step("verify autoload of bot", async function(done){
     try {
       const bot = new Bot({id: botData.id, server: botData.server});
-      expect(bot.id).to.be.equal(botData.id);
+      expect(bot.id).to.be.equal(botId);
       expect(bot.type).to.be.undefined;
       expect(bot.title).to.be.equal('');
       expect(bot.owner).to.be.undefined;
       expect(bot.server).to.be.equal(botData.server);
       expect(bot.loaded).to.be.equal(false);
-      
+
       const logged = await profile.connect(user, password, server);
-  
+
       when(() => bot.loaded, () => {
         try {
           expect(bot.title).to.be.equal(botData.title);
           expect(bot.type).to.be.equal(botData.type);
           expect(bot.server).to.be.equal(botData.server);
-          expect(bot.owner.user).to.be.equal(botData.owner);
+          expect(bot.owner.user).to.be.equal(botData.owner.user);
           expect(bot.visibility).to.be.equal(VISIBILITY_FRIENDS);
           done();
         } catch (e){
@@ -151,7 +151,7 @@ describe("bot", function() {
       done(e);
     }
   });
-  
+
   step("retrieve existing bot", async function(done){
     try {
       const data = await bot.load({id:botData.id, server:botData.server});
@@ -160,7 +160,7 @@ describe("bot", function() {
       expect(data.affiliations[0]).to.be.equal(friend);
       console.log("DATA:", data);
       expect(data.id).to.be.equal(botData.id);
-      
+
       await bot.publishContent(botData, 123, 'hello world!');
       await bot.publishContent(botData, 1234, 'hello world2!');
       let items = await bot.items(botData);
@@ -171,7 +171,7 @@ describe("bot", function() {
       await bot.removeItem(botData, 123);
       items = await bot.items(botData);
       expect(items.length).to.be.equal(0);
-  
+
       await bot.publishImage(botData, 1235, 'hello world url!');
       await bot.publishImage(botData, 1236, 'hello world url2!');
       await bot.publishImage(botData, 1237, 'hello world url2!');
@@ -185,28 +185,28 @@ describe("bot", function() {
       await bot.removeItem(botData, 1237);
       items = await bot.items(botData);
       expect(items.length).to.be.equal(0);
-  
+
       expect(data.title).to.be.equal(botData.title);
       expect(data.shortname).to.be.equal(botData.shortname);
       expect(data.server).to.be.equal(botData.server);
       expect(data.radius).to.be.equal(botData.radius);
       expect(data.description).to.be.equal(botData.description);
-      
+
       done();
     } catch (e) {
       done(e);
     }
   });
-  
+
   step("share bot headline", async function(done){
     try {
-      bot.share(botData, [friend, 'friends']);
+      bot.share(botData, [friend, 'friends'], 'headline');
       done();
     } catch (e){
       done(e);
     }
   });
-  
+
   step("retrieve list of own/following bots", async function(done){
     try {
       await botStore.start();
@@ -217,32 +217,17 @@ describe("bot", function() {
       done(e);
     }
   });
-  
-  // step("retrieve list of all bots", async function(done){
-  //   try {
-  //     const data = await bot.list(user, server);
-  //     console.log("DATA:", data.bots.length, data);
-  //     expect(data.bots.length > 0).to.be.true;
-  //     // remove bots
-  //     // for (let item of data.bots){
-  //     //   await bot.remove({id: item.id, server: item.server});
-  //     // }
-  //     done();
-  //   } catch (e){
-  //     done(e);
-  //   }
-  // });
-  //
+
   step("logout!", async function (done){
     await xmpp.disconnect(null);
     done();
   });
-  
+
   step("register/login friend and expect shared bot", async function(done){
     const data = testDataNew(12);
     const {user, password, server} = await xmpp.register(data.resource, data.provider_data);
     const logged = await xmpp.connect(user, password, server);
-    
+
     await xmpp.disconnect(null);
     done();
   });