+++ src/components/Home.js
@@ -9,7 +9,7 @@ export const HEIGHT = Dimensions.get('window').height;
 export const WIDTH = Dimensions.get('window').width;
 import NavBarCloseButton from './NavBarCloseButton';
 import assert from 'assert';
-import ActionButton from './ActionButton';
+import BotButton from './BotButton';
 import Chats from './ChatListView';
 import Map from './Map';
 import location from '../store/locationStore';
@@ -166,7 +166,7 @@ export default class Home extends React.Component {
         <Animated.View style={{flex:1, transform: [{translateY:this.state.top}]}}>
           <EventList ref="list" />
         </Animated.View>
-        <ActionButton/>
+        <BotButton/>
         <OwnMessageBar ref="alert" />
       </View>
     //   <FilterBar style={{paddingLeft:15*k, paddingRight:15*k}}