+++ CHANGELOG.md
@@ -11,8 +11,9 @@ If there are not many items, just list them sequentially.
 # Unreleased
 
 * Rename files to be unique. Remove unneeded files. 
+* Use 'new-id' method when creating bots. Set image permissions correctly. 
+* Fixes: Images not appearing on Bot (#323). 
 * Fix: Implement NEW Create Message CTA within Message Screen (#387)
-
 
 # 1.24.3 - 2017 March 3