+++ CHANGELOG.md
@@ -15,6 +15,7 @@ If there are not many items, just list them sequentially.
 * Use 'new-id' method when creating bots. Set image permissions correctly. 
 * Fixes: Images not appearing on Bot (#323). 
 * Fix: Implement NEW Create Message CTA within Message Screen (#387)
+* Implement NEW Create Bot CTA (#439)
 
 # 1.24.3 - 2017 March 3