+++ CHANGELOG.md
@@ -8,7 +8,7 @@ Suggested subheadings: `Added`, `Changed`, `Deprecated`, `Removed`, `Fixed`, `Se
 
 If there are not many items, just list them sequentially. 
 
-# Unreleased
+# 1.25.0 - 2017 March 9
 
 * Rename files to be unique. Remove unneeded files (#419). 
 * Fix: Bot Profile: Randomize Default Cover Photo (#406)